refactor: break up createTemplate
Break up `createTemplate` function from the VM action handler into
several smaller functions and move it and its associated functions into
a separate file.

The `createTemplate` function now only handles a few very basic tasks,
which are all implemented in their own separate functions. This makes
each function much easier to understand and it simplifies the code-paths
in the error cases.

In case the `createTemplate` function encounters an error in any of its
actions, it will clean up both the VirtualMachineTemplateVersion as well
as the VirtualMachineTemplate that it created before.

Moving this group of functions into a separate file makes it easier to
navigate the source code.
